# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Freak Hub API
  version: 0.1.0
  description: |
    HTTP contract between the Next.js front end and the Go API.

    This file is the source of truth: the TypeScript types the web consumes are
    generated from it (`pnpm contracts:generate`). Change the contract here
    first, then implement it on both sides.

    Authentication is a Clerk session JWT in `Authorization: Bearer <token>`,
    which the API verifies against Clerk's JWKS. The only unauthenticated
    endpoints are the health check and the Clerk webhook, and the latter
    authenticates with a Svix signature instead.

    ## Errors

    Every error response, without exception, is served as
    `application/problem+json` and shaped like `#/components/schemas/Error`
    (Problem Details, RFC 9457 — see
    [ADR-0014](../../docs/decisions/0014-sistema-de-estados-y-errores.md)).
    `code` is the stable, machine-readable identifier to branch on; `message`
    duplicates `detail` and exists only for backward compatibility with
    clients that predate this envelope.

    Every response, success or error, carries an `X-Request-ID` header — an
    opaque correlation id, also echoed in an error body's `correlation_id`.
    Include it when reporting an issue: it is what ties a response back to
    its server-side log line.

    Any endpoint below `/v1` and the webhook can also answer:

    - `405 Method Not Allowed` (`method_not_allowed`) for a verb the route
...
      uniformly.
...
servers:
...
paths:
...
  /v1/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: The group roster
      description: |
        Every member who has ever joined, ordered by member-since ascending
        (oldest first) and keyset-paginated per
        [ADR-0011](../../docs/decisions/0011-paginacion-por-cursor.md).
      parameters:
        - in: query
          name: limit
          description: Page size. Defaults to 25, must be between 1 and 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - in: query
          name: cursor
          description: Opaque cursor returned as `next_cursor` by a previous page.
          schema:
            type: string
      responses:
        "200":
          description: A page of the group roster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberPage"
        "400":
          description: The limit or the cursor is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/Timeout"
  /v1/invitations:
    get:
      operationId: listMyInvitations
      tags: [invitations]
      summary: Invitations sent by the current member
      responses:
        "200":
          description: The caller's invitations, newest first
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invitation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: |
            The session is valid but no member row exists yet, which happens
            while the `user.created` webhook is still in flight.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/Timeout"

    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Invite somebody
      description: |
        Any member can invite, with no quota. The API asks Clerk to send the
        email and records who invited whom.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
      responses:
        "201":
          description: The invitation was sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
        "400":
          description: The body is malformed or carries unknown fields
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: |
            The session is valid but no member row exists yet, which happens
            while the `user.created` webhook is still in flight.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: An invitation is already pending, or the email is already a member
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          description: The email is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/Timeout"
  /v1/invitations/group:
    get:
      operationId: listGroupInvitations
      tags: [invitations]
      summary: Every invitation the group has ever sent
      description: |
        Unlike `GET /v1/invitations`, which is scoped to the caller, this
        returns every invitation any member has sent, with who sent it, newest
        first and keyset-paginated per
        [ADR-0011](../../docs/decisions/0011-paginacion-por-cursor.md). It
        includes every status: the screen that shows pending guests can
        filter client-side, and a full trail is what lets two members notice
        they are about to invite the same address.
      parameters:
        - in: query
          name: limit
          description: Page size. Defaults to 25, must be between 1 and 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - in: query
          name: cursor
          description: Opaque cursor returned as `next_cursor` by a previous page.
          schema:
            type: string
      responses:
        "200":
          description: A page of the group's invitations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupInvitationPage"
        "400":
          description: The limit or the cursor is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: |
            The session is valid but no member row exists yet, which happens
            while the `user.created` webhook is still in flight.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/Timeout"
...
components:
  securitySchemes:
    clerkSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session token, obtained in the browser with `getToken()`.
  headers:
    RequestID:
      description: |
        The opaque correlation id for this request. Also present as
        `correlation_id` on an error body. Include it when reporting an
        issue.
      schema:
        type: string
  responses:
    Unauthorized:
      description: The session is missing, malformed or expired
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    PayloadTooLarge:
      description: The request body is larger than the endpoint accepts
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: A failure on our side. The detail is always generic.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: Clerk is rate limiting us or unreachable. Retryable.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Retry-After:
          description: Present when the delay before retrying is known.
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Timeout:
      description: The request did not finish inside the server's deadline. Retryable.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: |
        The single error envelope of the whole API (ADR-0014), served as
        `application/problem+json` and shaped after Problem Details (RFC
        9457). Branch on `code`, never on `title` or `detail`, which are
        human prose and may be reworded at any time. `message` duplicates
        `detail` and is deprecated: it exists only so a client written before
        ADR-0014 keeps working unmodified.
      required: [type, title, status, code, message, correlation_id]
      properties:
        type:
          type: string
          description: A URI identifying the problem type, `urn:freak-hub:problem:<code>`.
          examples: ["urn:freak-hub:problem:invitation_already_sent"]
        title:
          type: string
          description: A short, human-readable summary that does not change between occurrences of the same code.
        status:
          type: integer
          description: The HTTP status of this response, repeated in the body per RFC 9457.
        detail:
          type: string
          description: Prose specific to this occurrence, in Spanish.
        instance:
          type: string
          description: The request path that produced this error (never the query string).
        code:
          type: string
          description: |
            The stable, machine-readable identifier. This is what a client
            branches on. internal/api's contract parity test fails the build
            if this enum and the Go registry (httpx.Codes) ever disagree.
          examples: [invitation_already_sent, invalid_token, not_found]
          enum:
            - missing_token
            - invalid_token
            - unauthorized
            - unknown_identity
            - invalid_payload
            - not_found
            - internal_error
            - invalid_limit
            - invalid_cursor
            - no_profile_changes
            - name_too_long
            - username_invalid_length
            - username_numeric_only
            - username_taken
            - avatar_too_large
            - avatar_unsupported_type
            - invalid_email
            - invitation_already_sent
            - already_member
            - method_not_allowed
            - payload_too_large
            - request_timeout
            - upstream_unavailable
        message:
          type: string
          deprecated: true
          description: Same text as `detail`. Kept only for backward compatibility with clients written before ADR-0014.
        correlation_id:
          type: string
          description: Opaque request identifier, also returned on the `X-Request-ID` header. Include it when reporting an issue.
        retryable:
          type: boolean
          description: Whether retrying the exact same request has a reasonable chance of succeeding (429, 502, 503, 504-shaped conditions).
        retry_after:
          type: integer
          description: Suggested delay in seconds before retrying. Present only when `retryable` is true and the delay is actually known.
        field_errors:
          type: array
          description: Present only when the error unambiguously blames one or more request fields.
          items:
            type: object
            required: [field, code]
            properties:
              field:
                type: string
              code:
                type: string
    Member:
      type: object
      required: [id, clerk_user_id, username, display_name, created_at]
      properties:
        id:
          type: string
          format: uuid
        clerk_user_id:
          type: string
        username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        created_at:
          description: When this member joined ("member since").
          type: string
          format: date-time
    MemberPage:
      type: object
      description: |
        A keyset-paginated page (ADR-0011). `next_cursor` is `null` when
        nothing more remains to read.
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        next_cursor:
          type: [string, "null"]
    Invitation:
      type: object
      required: [id, email, status, inviter_id, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        status:
          type: string
          enum: [pending, accepted, revoked]
        inviter_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    InvitationInviter:
      type: object
      description: The public identity of the member who sent an invitation.
      required: [id, username, display_name]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
    GroupInvitation:
      type: object
      description: |
        An invitation as seen from the group's perspective: who was invited,
        by whom, and where it stands.
      required: [id, email, status, created_at, inviter]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        status:
          type: string
          enum: [pending, accepted, revoked]
        created_at:
          type: string
          format: date-time
        inviter:
          $ref: "#/components/schemas/InvitationInviter"
    GroupInvitationPage:
      type: object
      description: |
        A keyset-paginated page (ADR-0011). `next_cursor` is `null` when
        nothing more remains to read.
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GroupInvitation"
        next_cursor:
          type: [string, "null"]
security:
  - clerkSession: []